Don't auto-start local connections upon selection
There is apparently a race condition in building the UIs. The local
transport was starting up prematurely and causing the host list to cycle
multiple times at some point.

This change also makes it so that hosts quick-connected using the same
nickname as a previous host simply uses the previous host instead of
creating a new entry that can never be "reached"

git-svn-id: https://connectbot.googlecode.com/svn/trunk/connectbot@396 df292f66-193f-0410-a5fc-6d59da041ff2