Update to target API 15 (ICS 4.0.3).
This disables various compatibility features that we don't want, and
allows hardware acceleration to be enabled by default on ICS (explicitly
enabling it may have adverse effects on some HC devices).

Also, for API 14 and above, use the DeviceDefault theme rather than
explicitly using Holo, as this makes it look better with devices with
manufacturer-customised themes.