Handle git version differently to fix Eclipse builds.
Unfortunately Eclipse appears to be too dumb to generate the version and
build it correctly (even using an external build step it fails to
rebuild resources afterward), so go back to having a dummy version
checked into the repository. Run git update-index --assume-unchanged on
the file before updating the version, so that git doesn't consider the
file to have changed, instead.