Hopefully fix the majority of keyboard issues.
It seems that the majority of keyboard issues were caused by quirks
implemented for the convenience of the HTC Dream (shift+numbers
generating function keys on the hardware keyboard and ctrl+numbers
generating function keys on the software keyboard). Turn these into
settings, which are disabled by default unless the device is indeed
a HTC Dream.

Also, add a setting (defaulting to on) which determines whether the
volume keys change the font size. Disabling this restores normal
volume control functionality.