Scroll to the previously selected view asynchronously.
This avoids a race condition in ViewPager. See #301. Note that
this still looks a bit janky because the pager noticably scrolls
away and back to the right view. A bug has been filed with the
support library team to look into a better solution.

Also update the support library version.