Make emulated keys rearrangable
... by long click + drag and drop. This allows to move the most
freuently used keys to the left or to build functional groups,
individually. We can also add more keys, later, in such a way that they
still can be useul and at the same time don't affect users that don´t
need them. All this, without adding another preference option.

One drawback of this direct approach is that keys with auto-repeat
cannot be dragged. However, all other keys can be moved around them, so
that this drawback should be acceptable for a not frequently used
function.

The second drawback of not using another preference option menu might be
that the new functionally will remain an unfound easter egg for some.

Signed-off-by: Marc K. <morckx@gmail.com>