TelnetProtocolHandler Fix
This corrects intermittent issues with servers not receiving complete
TELOPT  sequence causing incorrect terminal and screen size detection.

Built Telnet Response Sequences in to an ArrayList<Byte> to transpose
back to a byte[] array so all response sequences are sent in a single
write() instead of sending each separate byte in an individual write().