Added Jenkinsfile
This is the first step in turning on the new Jenkins CI and turning down
the old one. Note that the pipeline isn't fully replicated from the old
Jenkins instance, but that will come soon.

This also enables anyone to run a pretty exact simulation of the Jenkins
build locally if any problem is encountered.