Fix text selection for certain font sizes
For certain font sizes (e.g. for 13 points), the overlay was using
slightly smaller font than the terminal bridge. It appears that the
overlay (i.e. TextView) is always rounding down the pixel size, whereas
TerminalBridge is rounding to the closest integer.

Copy the pixel size from the bridge to the overlay to ensure perfect
alignment between the two.

Remove TerminalBridge#setFontSize(), it's not used anymore. Add
TerminalBridge#setFontSizePx() instead.