Allow underscores in the hostmask
While the "rules" may say that _ is not allowed in host and/or
domainnames, the real worl reality is that they are widely used
and not breaking anything.  So embracing the robustness principle
and accepting the world as it is rather than how it should be, allow
the use of _ in host/domainnames.

Fixes: #188