Make onClick optional so repeated keys can handle
The repeated keys need to have their own touch handling so it will make
it able to handle the repeating of the keys when they're held down.
However, if we pass even an empty onClick handler to Surface, it will
consume the touch events. That's not what we want so we need to switch
to a non-touchable Surface.