Only display notification permission on connect
In order to be able to let the users get familiar with the app, we should
only show the permission request when it is needed.

Also if the user denies the notification permission, disable the
"persist connections" preference. That way the user can try to enable it
later in settings and get another notifications permission request.